WIP: polish GitHubClient
